+++ PetitMoteur3DCore/Core/Public/Mesh/chargeur.h
@@ -34,7 +34,7 @@ public:
 	virtual size_t GetNombreSommets() const = 0;
 	virtual size_t GetNombreIndex() const = 0;
 	virtual const void* GetIndexData()  const = 0;
-	virtual int GetNombreSubset() const = 0;
+	virtual size_t GetNombreSubset() const = 0;
 	virtual size_t  GetNombreMaterial() const = 0;
 	virtual CMaterial GetMaterial(int _i) const = 0;
